Split up header components
